requirements.txt
openai
python-dotenv
requests
pydantic
Mastodon.py
feedparser
replicate
fastapi
uvicorn[standard]
numpy
